Move previewLimitationsInfo to separate component
