+++ frontend/app-preview/src/views/LandingPage.test.tsx
@@ -39,46 +39,4 @@ describe('LandingPage', () => {
         expect(hidePreviewLimitationsTemporaryButton).toBeInTheDocument();
         expect(hidePreviewLimitationsForSessionButton).toBeInTheDocument();
     });
-
-    it('should close popover and not set value in session storage when hidePreviewLimitationsTemporaryButton is clicked', async () => {
-        renderWithMockStore()(<LandingPage variant={'preview'} />);
-
-        const user = userEvent.setup();
-
-        // Open popover
-        const previewLimitationsAlert = screen.getByText(textMock('preview.limitations_info'));
-        const alert = within(previewLimitationsAlert);
-        const hidePreviewLimitationsAlertButton = alert.getByRole('button');
-        await act(() => user.click(hidePreviewLimitationsAlertButton));
-        const hidePreviewLimitationsPopover = screen.getByText(textMock('session.reminder'));
-        expect(hidePreviewLimitationsPopover).toBeInTheDocument();
-        const hidePreviewLimitationsTemporaryButton = screen.getByRole('button', { name: textMock('session.do_show_again') });
-
-        // Click hide temporary button
-        await act(() => user.click(hidePreviewLimitationsTemporaryButton));
-
-        expect(hidePreviewLimitationsPopover).not.toBeInTheDocument();
-        expect(window.localStorage.getItem('showPreviewLimitationsInfo')).toBeNull();
-    });
-
-    it('should close popover and set value in session storage when hidePreviewLimitationsForSessionButton is clicked', async () => {
-        renderWithMockStore()(<LandingPage variant={'preview'} />);
-
-        const user = userEvent.setup();
-
-        // Open popover
-        const previewLimitationsAlert = screen.getByText(textMock('preview.limitations_info'));
-        const alert = within(previewLimitationsAlert);
-        const hidePreviewLimitationsAlertButton = alert.getByRole('button');
-        await act(() => user.click(hidePreviewLimitationsAlertButton));
-        const hidePreviewLimitationsPopover = screen.getByText(textMock('session.reminder'));
-        expect(hidePreviewLimitationsPopover).toBeInTheDocument();
-        const hidePreviewLimitationsForSessionButton = screen.getByRole('button', { name: textMock('session.dont_show_again') });
-
-        // Click hide forever button
-        await act(() => user.click(hidePreviewLimitationsForSessionButton));
-
-        expect(hidePreviewLimitationsPopover).not.toBeInTheDocument();
-        expect(window.localStorage.getItem('showPreviewLimitationsInfo')).toBe('false');
-    });
 });
