+++ frontend/app-preview/src/views/LandingPage.tsx
@@ -1,6 +1,5 @@
-import React, { useState } from 'react';
+import React from 'react';
 import classes from './LandingPage.module.css';
-import cn from 'classnames';
 import { useTranslation } from 'react-i18next';
 import { usePreviewConnection } from 'app-shared/providers/PreviewConnectionContext';
 import { useInstanceIdQuery, useRepoMetadataQuery, useUserQuery } from 'app-shared/hooks/queries';
@@ -14,11 +13,9 @@ import {
 } from '../components/AppBarConfig/AppPreviewBarConfig';
 import { appPreviewButtonActions } from '../components/AppBarConfig/AppPreviewBarConfig';
 import { AppPreviewSubMenu } from '../components/AppPreviewSubMenu';
-import { Alert, Button, LegacyPopover } from '@digdir/design-system-react';
-import { XMarkIcon } from '@navikt/aksel-icons';
 import { useStudioUrlParams } from 'app-shared/hooks/useStudioUrlParams';
 import { previewPage } from 'app-shared/api/paths';
-import { typedLocalStorage } from 'app-shared/utils/webStorage';
+import { PreviewLimitationsInfo } from 'app-shared/components/PreviewLimitationsInfo/PreviewLimitationsInfo';
 
 export interface LandingPageProps {
   variant?: AltinnHeaderVariant;
@@ -33,9 +30,6 @@ export const LandingPage = ({ variant = 'preview' }: LandingPageProps) => {
   const { data: user } = useUserQuery();
   const { data: repository } = useRepoMetadataQuery(org, app);
   const { data: instanceId } = useInstanceIdQuery(org, app);
-  const [openSaveChoiceInSession, setOpenShowSaveChoiceInSession] = useState<boolean>(false);
-  const showPreviewLimitationsInfoSession: boolean = typedLocalStorage.getItem('showPreviewLimitationsInfo');
-  const [showPreviewLimitationsInfo, setShowPreviewLimitationsInfo] = useState<boolean>(showPreviewLimitationsInfoSession ?? true);
   const [selectedLayoutSet, setSelectedLayoutSet] = useLocalStorage<string>('layoutSet/' + app, null);
   const [previewViewSize, setPreviewViewSize] = useLocalStorage<PreviewAsViewSize>(
     'viewSize',
@@ -52,17 +46,7 @@ export const LandingPage = ({ variant = 'preview' }: LandingPageProps) => {
     // might need to remove selected layout from local storage to make sure first page is selected
     window.location.reload();
   };
-
-  const handleHidePreviewLimitations = () => {
-    setShowPreviewLimitationsInfo(false);
-    setOpenShowSaveChoiceInSession(false);
-  };
-
-  const handleRememberChoiceForSession = () => {
-    typedLocalStorage.setItem('showPreviewLimitationsInfo', false);
-    handleHidePreviewLimitations();
-  };
-
+  
   if (previewConnection) {
     previewConnection.on('ReceiveMessage', function (message) {
       const frame = document.getElementById('app-frontend-react-iframe');
@@ -99,23 +83,8 @@ export const LandingPage = ({ variant = 'preview' }: LandingPageProps) => {
           }
         />
       </div>
-      <div className={classes.gridContainer}>
-      {showPreviewLimitationsInfo &&
-        <Alert severity='info' className={classes.previewLimitationsInfo}>
-          <div className={classes.alert}>
-            {t('preview.limitations_info')}
-            <LegacyPopover
-                trigger={<Button onClick={() => setOpenShowSaveChoiceInSession(!openSaveChoiceInSession)} size='small' variant='tertiary' icon={<XMarkIcon />}/>}
-                open={openSaveChoiceInSession}
-            >
-              <div className={classes.grid}>
-                <p className={classes.message}>{t('session.reminder')}</p>
-                <Button className={cn(classes.yes, classes.button)}  onClick={handleHidePreviewLimitations} size='small' variant='secondary'>{t('session.do_show_again')}</Button>
-                <Button className={cn(classes.no, classes.button)} onClick={handleRememberChoiceForSession} size='small' variant='secondary'>{t('session.dont_show_again')}</Button>
-              </div>
-            </LegacyPopover>
-          </div>
-        </Alert>}
+      <div className={classes.previewArea}>
+        <PreviewLimitationsInfo />
         <div className={classes.iframeContainer}>
           <iframe
             title={t('preview.iframe_title')}
