+++ frontend/packages/shared/src/components/PreviewLimitationsInfo/PreviewLimitationsInfo.module.css
@@ -1,0 +1,16 @@
+.alert {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    font-family: var(--font-family);
+}
+
+.popoverContent {
+    display: grid;
+    grid-template-columns: min-content min-content;
+    gap: .5rem;
+}
+.message { grid-area: 1 / 1 / 2 / 3; margin: 0 }
+.popoverContent .yesButton { grid-area: 2 / 1 / 3 / 2 }
+.popoverContent .noButton { grid-area: 2 / 2 / 3 / 3 }
+.popoverContent .button { white-space: nowrap }
